Move `FilesItemViewModel` to it's own file
