+++ WireMessaging/Sources/WireMessagingUI/WireCells/Components/Files/FilesViewModel.swift
@@ -158,117 +158,3 @@ package final class FilesViewModel: ObservableObject {
     }
 
 }
-
-@MainActor
-/// A view model for a single item in the `FilesView`.
-///
-/// A view model is needed as the item is _live_ - it can be updated remotely, it's file downloaded locally, and so on.
-/// A locally downloaded file may also become out of date and need to be re-downloaded. Using a view model allows us to
-/// have granular subscriptions to events that are cancelled when the view is no longer in view.
-final class FilesItemViewModel: ObservableObject {
-
-    private let nodeID: UUID
-    private var cancellables = Set<AnyCancellable>()
-
-    @Published private var asset: WireCellsLocalAsset?
-
-    let fileName: String
-    let subtitle: String?
-    let icon: FileIcon
-    let localAssetRepository: any WireCellsLocalAssetRepositoryProtocol
-
-    init(
-        item: FilesViewItem,
-        localAssetRepository: any WireCellsLocalAssetRepositoryProtocol,
-        locale: Locale = .autoupdatingCurrent,
-        calendar: Calendar = .autoupdatingCurrent,
-        timeZone: TimeZone = .autoupdatingCurrent
-    ) {
-        self.nodeID = item.id
-        self.fileName = item.filename
-        self.subtitle = Self.subtitle(from: item, locale: locale, calendar: calendar, timeZone: timeZone)
-        self.icon = item.icon
-        self.localAssetRepository = localAssetRepository
-
-        localAssetRepository.observeAsset(nodeID: nodeID).sink { [self] asset in
-            self.asset = asset
-        }.store(in: &cancellables)
-    }
-
-    var isDownloaded: Bool {
-        switch asset?.downloadState {
-        case .downloaded:
-            return true
-        default:
-            return false
-        }
-    }
-
-    var isDownloadOptionAvailable: Bool {
-        switch asset?.downloadState {
-        case .downloaded:
-            return true
-        default:
-            return false
-        }
-    }
-
-    var isDownloadOptionDisabled: Bool {
-        switch asset?.downloadState {
-        case .downloading:
-            return true
-        default:
-            return false
-        }
-    }
-
-    var progress: Double? {
-        switch asset?.downloadState {
-        case .downloading(let progress):
-            return progress
-        case .failed:
-            return 100 // We show a full red progress bar on failure
-        default:
-            return nil
-        }
-    }
-
-    var showErrorState: Bool {
-        switch asset?.downloadState {
-        case .failed:
-            return true
-        default:
-            return false
-        }
-    }
-
-    func download() async {
-        // Ignore errors as these will be reported via the `asset` publisher.
-        try? await localAssetRepository.downloadAsset(nodeID: nodeID)
-    }
-
-    private static func subtitle(
-        from item: FilesViewItem,
-        locale: Locale,
-        calendar: Calendar,
-        timeZone: TimeZone
-    ) -> String? {
-        let modifiedAt = item.modifiedAt.map { date in
-            let style = Date.FormatStyle(
-                date: .abbreviated,
-                time: .shortened,
-                locale: locale,
-                calendar: calendar,
-                timeZone: timeZone,
-                capitalizationContext: .beginningOfSentence
-            )
-            return date.formatted(style)
-        }
-        return if let modifiedAt, let ownedBy = item.ownedBy {
-            L10n.Localizable.Conversation.WireCells.Files.Item.subtitle(modifiedAt, ownedBy)
-        } else {
-            [modifiedAt, item.ownedBy].compactMap(\.self).first
-        }
-    }
-
-}
